move svgs
