+++ frontend/.prettierignore
@@ -4,3 +4,4 @@ build/
 dist/
 out/
 public/
+assets/
