+++ frontend/src/components/ExportButton.tsx
@@ -1,5 +1,6 @@
 import styles from "./CSVButton.module.css";
 import Image from "next/image";
+import uploadBlue from "@/assets/upload_blue.svg";
 
 export default function ExportButton() {
   return (
@@ -7,7 +8,7 @@ export default function ExportButton() {
       <div className={styles.exportText}>Export CSV</div>
       <span className={styles.ic_container}>
         <Image
-          src="/upload_blue.svg"
+          src={uploadBlue}
           alt="Upload logo"
           className={styles.uploadIcon}
           width={13.333}
