+++ frontend/src/components/ImportButton.tsx
@@ -1,5 +1,6 @@
 import styles from "./CSVButton.module.css";
 import Image from "next/image";
+import uploadIcon from "@/assets/upload.svg";
 
 type ImportButtonProps = {
   onClick: () => void;
@@ -10,7 +11,7 @@ export default function ImportButton({ onClick }: ImportButtonProps) {
     <button type="button" className={styles.importButton} onClick={onClick}>
       <span className={styles.ic_container}>
         <Image
-          src="/upload.svg"
+          src={uploadIcon}
           alt="Upload logo"
           className={styles.uploadIcon}
           width={13.333}
