+++ frontend/src/components/PageBar.tsx
@@ -2,6 +2,7 @@
 
 import styles from "./PageBar.module.css";
 import Image from "next/image";
+import caretIcon from "@/assets/caret.svg";
 
 interface PageBarProps {
   totalItems: number;
@@ -50,7 +51,7 @@ export default function PageBar({
         <div className={styles.previous} onClick={() => onPageChange(Math.max(1, currentPage - 1))}>
           <div className={styles.ic_container}>
             <Image
-              src="/caret.svg"
+              src={caretIcon}
               alt="Caret Left"
               className={styles.caretIconLeft}
               width={24}
@@ -77,7 +78,7 @@ export default function PageBar({
         >
           <div className={styles.ic_container}>
             <Image
-              src="/caret.svg"
+              src={caretIcon}
               alt="Caret Right"
               className={styles.caretIconRight}
               width={24}
