+++ frontend/src/components/SearchBar.tsx
@@ -3,6 +3,8 @@
 import styles from "./SearchBar.module.css";
 import { useEffect, useRef, useState } from "react";
 import Image from "next/image";
+import unionIcon from "@/assets/Union.svg";
+import checkboxIcon from "@/assets/Checkbox.svg";
 
 interface SearchBarProps {
   search: string;
@@ -94,7 +96,7 @@ export default function SearchBar({
         <div className={styles.inputField}>
           <span className={styles.ic_search}>
             <Image
-              src="/Union.svg"
+              src={unionIcon}
               alt="Union logo"
               className={styles.union}
               width={24}
@@ -127,7 +129,7 @@ export default function SearchBar({
                 >
                   {selected.has(item) && (
                     <Image
-                      src="/Checkbox.svg"
+                      src={checkboxIcon}
                       alt="checked"
                       className={styles.checkIcon}
                       width={16}
@@ -149,7 +151,7 @@ export default function SearchBar({
         <div className={styles.inputField}>
           <span className={styles.ic_search}>
             <Image
-              src="/Union.svg"
+              src={unionIcon}
               alt="Union logo"
               className={styles.union}
               width={24}
